Remove e2e package. Move Cypress into main repo
