+++ .gitignore
@@ -19,3 +19,8 @@ coverage
 
 # Work-in-progress folder
 wip
+
+# Cypress
+downloads
+screenshots
+videos
