+++ bundle/cypress/e2e/parallel-5/right-slot.cy.ts
