+++ bundle/cypress/fixtures/api/regular-stories.ts
