+++ bundle/cypress/fixtures/pages/fronts.ts
