+++ bundle/cypress/snapshot/snapshot-article-standard.cy.ts
