+++ bundle/cypress/support/e2e.ts
@@ -64,7 +64,7 @@ beforeEach(() => {
 
 // Hide fetch/XHR requests
 const app = window.top;
-if (!app.document.head.querySelector('[data-hide-command-log-request]')) {
+if (app && !app.document.head.querySelector('[data-hide-command-log-request]')) {
 	const style = app.document.createElement('style');
 	style.innerHTML =
 		'.command-name-request, .command-name-xhr { display: none }';
