+++ bundle/package.json
@@ -23,6 +23,10 @@
 		"clean": "rm -rf dist",
 		"compile:core:common": "tsc --project ./tsconfig.core.json --outDir ./dist/cjs --module CommonJS",
 		"compile:core:esm": "tsc --project ./tsconfig.core.json --outDir ./dist/esm",
+		"e2e:open": "DEBUG=cypress:* cypress open --e2e --browser=chrome",
+		"e2e:run": "cypress run --spec 'cypress/e2e/**/*'",
+		"e2e:visual:open": "cypress open --e2e --browser=chrome --config specPattern='cypress/snapshot/**/*'",
+		"e2e:visual:run": "percy exec -- cypress run --config specPattern='cypress/snapshot/**/*'",
 		"lint": "eslint .",
 		"precommit:lint": "lint-staged",
 		"prepush:test": "jest --verbose --runInBand --onlyChanged",
@@ -48,8 +52,11 @@
 		"@babel/runtime": "^7.11.2",
 		"@commitlint/cli": "^17.0.3",
 		"@commitlint/config-conventional": "^17.0.0",
+		"@cypress/webpack-preprocessor": "^5.17.1",
 		"@guardian/eslint-config-typescript": "^5.0.0",
 		"@guardian/prettier": "^3.0.0",
+		"@percy/cli": "^1.24.2",
+		"@percy/cypress": "^3.1.2",
 		"@semantic-release/github": "^8.0.2",
 		"@types/google.analytics": "^0.0.42",
 		"@types/googletag": "~1.1.3",
@@ -67,11 +74,13 @@
 		"commitizen": "^4.2.4",
 		"conventional-changelog-conventionalcommits": "^5.0.0",
 		"csstype": "^3.1.1",
+		"cypress": "^12.13.0",
 		"cz-conventional-changelog": "^3.3.0",
 		"eslint": "^8.4.1",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-import-resolver-alias": "^1.1.2",
 		"eslint-import-resolver-typescript": "^3.5.3",
+		"eslint-plugin-cypress": "^2.13.3",
 		"eslint-plugin-eslint-comments": "^3.2.0",
 		"eslint-plugin-import": "^2.25.3",
 		"eslint-plugin-jest": "^27.2.1",
