--- e2e/.gitignore
@@ -1,2 +1,0 @@
-screenshots
-videos
